Crate solana_sysvar_id

Access to special accounts with dynamically-updated data.

Sysvars are special accounts that contain dynamically-updated data about the network cluster, the blockchain history, and the executing transaction. Each sysvar is defined in its own crate. The clock, epoch_schedule, instructions, and rent sysvars are most useful to on-chain programs.

All sysvar accounts are owned by the account identified by solana_sysvar::ID.

For more details see the Solana documentation on sysvars.

Macros

declare_deprecated_id

Re-export types required for macros Same as declare_id except that it reports that this ID has been deprecated.

declare_deprecated_sysvar_id

Same as declare_sysvar_id except that it reports that this ID has been deprecated.

declare_id

Re-export types required for macros Convenience macro to declare a static public key and functions to interact with it.

declare_sysvar_id

Declares an ID that implements SysvarId.

impl_deprecated_sysvar_id

Implements SysvarId for a module that already uses `declare_deprecated_id``

impl_sysvar_id

Implements SysvarId for a module that already uses `declare_id``

Structs

Pubkey

Re-export types required for macros The address of a Solana account.

Constants

ID

Re-export types required for macros The const program ID.

Traits

SysvarId

A type that holds sysvar data and has an associated sysvar Pubkey.

Functions

check_id

Re-export types required for macros Returns true if given pubkey is the program ID.

id

Re-export types required for macros Returns the program ID.